// src/api/models/user-im.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Resume Builder API
 * An API for Online Resume Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UserIM
 */
export interface UserIM {
    /**
     * 
     * @type {string}
     * @memberof UserIM
     */
    firstName: string;
    /**
     * 
     * @type {string}
     * @memberof UserIM
     */
    lastName: string;
    /**
     * 
     * @type {string}
     * @memberof UserIM
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof UserIM
     */
    password: string;
}