// src/api/models/user-lm.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Resume Builder API
 * An API for Online Resume Builder
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UserLM
 */
export interface UserLM {
    /**
     * 
     * @type {string}
     * @memberof UserLM
     */
    email: string;
    /**
     * 
     * @type {string}
     * @memberof UserLM
     */
    password: string;
}